UI-25 types
